vectorsentinel/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vectorsentinel"
version = "0.1.0"
description = "Stop RAG hallucinations before the LLM is called. A drop-in Python library that adds 'I don't know' intelligence to any vector search pipeline."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [{ name = "VectorSentinel Contributors" }]
keywords = [
    "rag",
    "hallucination",
    "vector-search",
    "knn",
    "out-of-distribution",
    "confidence",
    "llm",
    "retrieval",
    "deterministic",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = ["numpy>=1.24"]

[project.optional-dependencies]
faiss = ["faiss-cpu>=1.7"]
chroma = ["chromadb>=0.4"]
langchain = ["langchain-core>=0.1"]
llamaindex = ["llama-index-core>=0.10"]
server = ["fastapi>=0.100", "uvicorn>=0.20"]
all = [
    "faiss-cpu>=1.7",
    "chromadb>=0.4",
    "langchain-core>=0.1",
    "llama-index-core>=0.10",
    "fastapi>=0.100",
    "uvicorn>=0.20",
]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1"]

[project.scripts]
sentinel = "vectorsentinel.cli:main"

[project.urls]
Homepage = "https://github.com/vectorsentinel/vectorsentinel"
Documentation = "https://vectorsentinel.dev"
Repository = "https://github.com/vectorsentinel/vectorsentinel"
Issues = "https://github.com/vectorsentinel/vectorsentinel/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/vectorsentinel"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
